+++ src/Controller/AssociationController.php
@@ -12,7 +12,7 @@ use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
 /**
- * @Route ("/main/association")
+ * @Route ("/association")
  */
 class AssociationController extends AbstractController
 {
