+++ src/Controller/CategoryController.php
@@ -11,7 +11,7 @@ use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
 /**
- * @Route("/main/category")
+ * @Route("/category")
  */
 class CategoryController extends AbstractController
 {
