+++ src/Controller/CommentController.php
@@ -11,7 +11,7 @@ use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
 /**
- * @Route("/main/comment")
+ * @Route("/comment")
  */
 class CommentController extends AbstractController
 {
