+++ src/Controller/GalleryController.php
@@ -11,7 +11,7 @@ use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
 /**
- * @Route("/admin/gallery")
+ * @Route("/gallery")
  */
 class GalleryController extends AbstractController
 {
