+++ src/Controller/ShopController.php
@@ -2,28 +2,93 @@
 
 namespace App\Controller;
 
+use App\Entity\Product;
+use App\Form\ProductType;
+use App\Repository\ProductRepository;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
+/**
+ * @Route ("/product")
+ */
 class ShopController extends AbstractController
 {
     /**
-     * @Route("/shop", name="shop")
+     * @Route("/", name="product_index", methods={"GET"})
      */
-    public function shop()
+    public function index(ProductRepository $productRepository): Response
     {
-        return $this->render('shop/shop.html.twig', [
-            'controller_name' => 'ShopController',
+        return $this->render('main/product/index.html.twig', [
+            'products' => $productRepository->findAll(),
         ]);
     }
 
     /**
-     * @Route("/shop/item/{id}", name="item")
+     * @Route("/new", name="product_new", methods={"GET","POST"})
      */
-    public function item($id)
+    public function new(Request $request): Response
     {
-        return $this->render('shop/item.html.twig', [
-            'id' => $id,
+        $product = new Product();
+        $form = $this->createForm(ProductType::class, $product);
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+            $entityManager = $this->getDoctrine()->getManager();
+            $entityManager->persist($product);
+            $entityManager->flush();
+
+            return $this->redirectToRoute('product_index');
+        }
+
+        return $this->render('main/product/new.html.twig', [
+            'product' => $product,
+            'form' => $form->createView(),
         ]);
+    }
+
+    /**
+     * @Route("/{id}", name="product_show", methods={"GET"})
+     */
+    public function show(Product $product): Response
+    {
+        return $this->render('main/product/show.html.twig', [
+            'product' => $product,
+        ]);
+    }
+
+    /**
+     * @Route("/{id}/edit", name="product_edit", methods={"GET","POST"})
+     */
+    public function edit(Request $request, Product $product): Response
+    {
+        $form = $this->createForm(ProductType::class, $product);
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+            $this->getDoctrine()->getManager()->flush();
+
+            return $this->redirectToRoute('product_index');
+        }
+
+        return $this->render('main/product/edit.html.twig', [
+            'product' => $product,
+            'form' => $form->createView(),
+        ]);
+    }
+
+    /**
+     * @Route("/{id}", name="product_delete", methods={"DELETE"})
+     */
+    public function delete(Request $request, Product $product): Response
+    {
+        if ($this->isCsrfTokenValid('delete'.$product->getId(), $request->request->get('_token'))) {
+            $entityManager = $this->getDoctrine()->getManager();
+            $entityManager->remove($product);
+            $entityManager->flush();
+        }
+
+        return $this->redirectToRoute('product_index');
     }
 }
